+++ src/batcher.rs
@@ -47,6 +47,7 @@ pub struct Batcher<F>
 where
     F: Fetcher,
 {
+    label: String,
     cache_store: CacheStore<F::Key, F::Value>,
     _fetch_task: Arc<tokio::task::JoinHandle<()>>,
     fetch_request_tx: tokio::sync::mpsc::Sender<FetchRequest<F::Key>>,
@@ -121,6 +122,7 @@ where
             fetcher,
             delay_duration: tokio::time::Duration::from_millis(10),
             eager_batch_size: Some(100),
+            label: format!("unlabeled-batcher"),
         }
     }
 
@@ -143,14 +145,18 @@ where
     /// detailed loading semantics.
     pub async fn load_many(&self, keys: &[F::Key]) -> Result<Vec<F::Value>, LoadError> {
         log::trace!(
-            "Looking up a batch of keys ({num_keys} key(s))",
-            num_keys=keys.len()
+            "[{label}/load_many] Looking up a batch of keys ({num_keys} key(s))",
+            label=self.label,
+            num_keys=keys.len(),
         );
         let mut cache_lookup = CacheLookup::new(keys.to_vec());
 
         match cache_lookup.lookup(&self.cache_store).await {
             CacheLookupState::Done(result) => {
-                log::trace!("All keys have already been looked up");
+                log::trace!(
+                    "[{label}/load_many] All keys have already been looked up",
+                    label=self.label,
+                );
                 return result;
             }
             CacheLookupState::Pending => {}
@@ -161,8 +167,9 @@ where
         let (result_tx, result_rx) = tokio::sync::oneshot::channel();
 
         log::debug!(
-            "Sending a batch of keys to fetch ({num_keys} key(s) still pending)",
-            num_keys=pending_keys.len()
+            "[{label}/load_many] Sending a batch of keys to fetch ({num_keys} key(s) still pending)",
+            label=self.label,
+            num_keys=pending_keys.len(),
         );
         let fetch_request = FetchRequest {
             keys: pending_keys,
@@ -175,24 +182,41 @@ where
 
         match result_rx.await {
             Ok(Ok(())) => {
-                log::debug!("Fetch response returned successfully");
+                log::debug!(
+                    "[{label}/load_many] Fetch response returned successfully",
+                    label=self.label,
+                );
             }
             Ok(Err(fetch_error)) => {
-                log::info!("Error message returned while fetching keys: {}", fetch_error);
+                log::info!(
+                    "[{label}/load_many] Error message returned while fetching keys: {error}",
+                    label=self.label,
+                    error=fetch_error,
+                );
                 return Err(LoadError::FetchError(fetch_error));
             }
             Err(recv_error) => {
-                panic!("Batch result channel hung up with error: {}", recv_error);
+                panic!(
+                    "Batch result channel for batcher {label} hung up with error: {error}",
+                    label=self.label,
+                    error=recv_error,
+                );
             }
         }
 
         match cache_lookup.lookup(&self.cache_store).await {
             CacheLookupState::Done(result) => {
-                log::trace!("All keys have now been looked up");
+                log::trace!(
+                    "[{label}/load_many] All keys have now been looked up",
+                    label=self.label,
+                );
                 return result;
             }
             CacheLookupState::Pending => {
-                panic!("Batch result is still pending after result channel was sent");
+                panic!(
+                    "Batch result for batcher {label} is still pending after result channel was sent",
+                    label=self.label,
+                );
             }
         }
     }
@@ -207,6 +231,7 @@ where
             cache_store: self.cache_store.clone(),
             _fetch_task: self._fetch_task.clone(),
             fetch_request_tx: self.fetch_request_tx.clone(),
+            label: self.label.clone(),
         }
     }
 }
@@ -220,6 +245,7 @@ where
     fetcher: F,
     delay_duration: tokio::time::Duration,
     eager_batch_size: Option<usize>,
+    label: String,
 }
 
 impl<F> BatcherBuilder<F>
@@ -249,12 +275,20 @@ where
         self
     }
 
+    /// Set a label for the `Batcher`. This is only used to improve diagnostic
+    /// messages, such as logs.
+    pub fn label(mut self, label: impl Into<String>) -> Self {
+        self.label = label.into();
+        self
+    }
+
     /// Create and return a `Batcher` with the given options.
     pub fn build(self) -> Batcher<F> {
         let cache_store = CacheStore::new();
 
         let (fetch_request_tx, mut fetch_request_rx) =
             tokio::sync::mpsc::channel::<FetchRequest<F::Key>>(1);
+        let label = self.label.clone();
 
         let fetch_task = tokio::spawn({
             let cache_store = cache_store.clone();
@@ -264,7 +298,10 @@ where
                     let mut pending_keys = HashSet::new();
                     let mut result_txs = vec![];
 
-                    log::debug!("Waiting for keys to fetch...");
+                    log::debug!(
+                        "[{label}/fetch_task] Waiting for keys to fetch...",
+                        label=self.label,
+                    );
                     match fetch_request_rx.recv().await {
                         Some(fetch_request) => {
                             for key in fetch_request.keys {
@@ -287,8 +324,9 @@ where
                         if should_run_batch_now {
                             // We have enough keys already, so don't wait for more
                             log::trace!(
-                                "Ready to fetch keys ({num_keys} key(s) pending)",
-                                num_keys=pending_keys.len()
+                                "[{label}/fetch_task] Ready to fetch keys ({num_keys} key(s) pending)",
+                                label=self.label,
+                                num_keys=pending_keys.len(),
                             );
                             break 'wait_for_more_keys;
                         }
@@ -303,7 +341,8 @@ where
                                         }
 
                                         log::trace!(
-                                            "Fetch request received ({num_keys} total key(s) pending)",
+                                            "[{label}/fetch_task] Fetch request received ({num_keys} total key(s) pending)",
+                                            label=self.label,
                                             num_keys=pending_keys.len()
                                         );
                                         result_txs.push(fetch_request.result_tx);
@@ -311,8 +350,9 @@ where
                                     None => {
                                         // Fetch queue closed, so we're done waiting for keys
                                         log::debug!(
-                                            "Fetch queue closed ({num_keys} key(s) pending)",
-                                            num_keys=pending_keys.len()
+                                            "[{label}/fetch_task] Fetch queue closed ({num_keys} key(s) pending)",
+                                            label=self.label,
+                                            num_keys=pending_keys.len(),
                                         );
                                         break 'wait_for_more_keys;
                                     }
@@ -322,8 +362,9 @@ where
                             _ = &mut delay => {
                                 // Reached delay, so we're done waiting for keys
                                 log::trace!(
-                                    "Delay reached, ready to send keys ({num_keys} key(s) pending)",
-                                    num_keys=pending_keys.len()
+                                    "[{label}/fetch_task] Delay reached, ready to send keys ({num_keys} key(s) pending)",
+                                    label=self.label,
+                                    num_keys=pending_keys.len(),
                                 );
                                 break 'wait_for_more_keys;
                             }
@@ -334,8 +375,9 @@ where
                         let mut cache = cache_store.as_cache();
 
                         log::debug!(
-                            "Fetching keys ({num_keys} key(s) to fetch)",
-                            num_keys=pending_keys.len()
+                            "[{label}/fetch_task] Fetching keys ({num_keys} key(s) to fetch)",
+                            label=self.label,
+                            num_keys=pending_keys.len(),
                         );
                         let pending_keys: Vec<_> = pending_keys.into_iter().collect();
                         let result = self
@@ -360,6 +402,7 @@ where
         });
 
         Batcher {
+            label,
             cache_store,
             _fetch_task: Arc::new(fetch_task),
             fetch_request_tx,